8292006: Move thread accessor classes to threadJavaClasses.hpp
