+++ src/hotspot/share/runtime/continuation.cpp
@@ -33,6 +33,7 @@
 #include "runtime/interfaceSupport.inline.hpp"
 #include "runtime/javaThread.inline.hpp"
 #include "runtime/osThread.hpp"
+#include "runtime/threadJavaClasses.inline.hpp"
 #include "runtime/vframe.inline.hpp"
 #include "runtime/vframe_hp.hpp"
