+++ src/hotspot/share/runtime/thread.cpp
@@ -25,8 +25,6 @@
 
 #include "precompiled.hpp"
 #include "jvm.h"
-#include "classfile/javaClasses.hpp"
-#include "classfile/javaThreadStatus.hpp"
 #include "gc/shared/barrierSet.hpp"
 #include "jfr/jfrEvents.hpp"
 #include "jvmtifiles/jvmtiEnv.hpp"
@@ -38,12 +36,14 @@
 #include "runtime/atomic.hpp"
 #include "runtime/handles.inline.hpp"
 #include "runtime/javaThread.inline.hpp"
+#include "runtime/javaThreadStatus.hpp"
 #include "runtime/nonJavaThread.hpp"
 #include "runtime/orderAccess.hpp"
 #include "runtime/osThread.hpp"
 #include "runtime/safepoint.hpp"
 #include "runtime/safepointMechanism.inline.hpp"
 #include "runtime/thread.inline.hpp"
+#include "runtime/threadJavaClasses.hpp"
 #include "runtime/threadSMR.inline.hpp"
 #include "services/memTracker.hpp"
 #include "utilities/macros.hpp"
