+++ src/hotspot/share/utilities/exceptions.cpp
@@ -40,6 +40,7 @@
 #include "runtime/javaThread.hpp"
 #include "runtime/os.hpp"
 #include "runtime/threadCritical.hpp"
+#include "runtime/threadJavaClasses.hpp"
 #include "runtime/atomic.hpp"
 #include "utilities/events.hpp"
 #include "utilities/exceptions.hpp"
